+++ src/common/components/batch-allocation/AllocationToolsMenuSection/AllocationToolsMenuSection.js
@@ -14,7 +14,6 @@ import {
 export const AllocationToolsMenuSection = ({
   disabled,
   onBatchAllocate,
-  onBatchAllocationLogs,
   onDownloadAllocationWorksheet,
   onUploadAllocationWorksheet,
 }) => {
@@ -69,22 +68,6 @@ export const AllocationToolsMenuSection = ({
           </Icon>
         </Button>
       </IfPermission>
-
-      <IfPermission perm="ui-finance.fund-update-logs.view">
-        <Button
-          buttonStyle="dropdownItem"
-          onClick={onBatchAllocationLogs}
-          disabled={disabled}
-        >
-          <Icon
-            size="small"
-            icon="report"
-          >
-            <FormattedMessage id="ui-finance.actions.allocations.batch.logs" />
-          </Icon>
-        </Button>
-      </IfPermission>
-
     </MenuSection>
   );
 };
@@ -92,7 +75,6 @@ export const AllocationToolsMenuSection = ({
 AllocationToolsMenuSection.propTypes = {
   disabled: PropTypes.bool,
   onBatchAllocate: PropTypes.func.isRequired,
-  onBatchAllocationLogs: PropTypes.func.isRequired,
   onDownloadAllocationWorksheet: PropTypes.func.isRequired,
   onUploadAllocationWorksheet: PropTypes.func.isRequired,
 };